NXBT-3306/NXBT-3310: move review tests where they belong
